requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
alembic==1.12.0
pymysql==1.0.2
python-ldap==3.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.8.2
pydantic-settings==2.0.3
python-multipart==0.0.5
python-dotenv==0.19.0
redis==4.3.4
email-validator==2.0.0 